Move all color and style management into a single class for easy maintenance
